% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{icbm}
\alias{icbm}
\title{ICBM-DTI-81 White Matter Atlas}
\usage{
icbm()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Brain atlas for the ICBM DTI-81 white matter labels with 26
tracts. Contains 3D mesh geometry for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
No 2D polygon geometry is available for this atlas.
}
\examples{
icbm()
}
\references{
Mori S, et al. (2005). MRI Atlas of Human White
Matter. Elsevier. ISBN: 978-0444517418
}
\seealso{
Other ggseg_atlases: 
\code{\link{jhu_labels}()},
\code{\link{jhu_tracts}()}

Other tract_atlases: 
\code{\link{jhu_labels}()},
\code{\link{jhu_tracts}()}
}
\concept{ggseg_atlases}
\concept{tract_atlases}
