% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{jhu}
\alias{jhu}
\title{JHU White Matter Atlas}
\usage{
jhu()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Brain atlas for the JHU white matter tractography with 11 tracts
per hemisphere. Contains both 2D polygon geometry for
\code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and 3D mesh geometry for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
jhu()
}
\references{
Hua et al. (2008) Tract probability maps in stereotaxic spaces:
analysis of white matter anatomy and tract-specific quantification.
NeuroImage, 39(1):336-347
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2724595/}{PubMed})
}
\concept{ggseg_atlases}
