% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{jhu_labels}
\alias{jhu_labels}
\title{JHU ICBM-DTI-81 White Matter Labels}
\usage{
jhu_labels()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
ICBM-DTI-81 white matter labels atlas with 48 regions
derived from a probabilistic atlas of 81 normal subjects.
Contains 3D mesh geometry for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
jhu_labels()
}
\references{
Mori S, et al. (2005). MRI Atlas of Human White
Matter. Elsevier. ISBN: 978-0444517418

Wakana S, et al. (2007). Reproducibility of quantitative
tractography methods applied to cerebral white matter.
\emph{NeuroImage}, 36(3):630-644.
\doi{10.1016/j.neuroimage.2007.02.049}
}
\seealso{
Other ggseg_atlases: 
\code{\link{icbm}()},
\code{\link{jhu_tracts}()}

Other tract_atlases: 
\code{\link{icbm}()},
\code{\link{jhu_tracts}()}
}
\concept{ggseg_atlases}
\concept{tract_atlases}
