% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegJHU-package.R
\docType{package}
\name{ggsegJHU-package}
\alias{ggsegJHU}
\alias{ggsegJHU-package}
\title{ggsegJHU: JHU White Matter Atlases for the 'ggseg' Ecosystem}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

JHU and ICBM white matter atlases for the 'ggseg' ecosystem. Includes the JHU probabilistic tract atlas (Hua et al. 2008), built from its probabilistic maps as a tract atlas with a centerline per tract, and the ICBM-DTI-81 white matter labels (Mori et al. 2005). Both are coregistered to the FreeSurfer 'cvs_avg35_inMNI152' template so 2D slices show cortical reference outlines, and ship as 'ggseg_atlas' objects with 3D geometry for use with 'ggseg3d'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegJHU}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegJHU/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Piñeiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

}
\keyword{internal}
