% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu.R
\name{jhu}
\alias{jhu}
\title{JHU White Matter Tract Atlas}
\usage{
jhu()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Probabilistic white matter tract atlas based on the
Johns Hopkins University DTI-based tractography.
}
\examples{
jhu()
}
\references{
Hua et al. (2008) Tract probability maps in stereotaxic spaces:
analysis of white matter anatomy and tract-specific quantification.
NeuroImage, 39(1):336-347
\doi{10.1016/j.neuroimage.2007.07.053}
}
\concept{ggseg_atlases}
