% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_tracts.R
\name{jhu_tracts}
\alias{jhu_tracts}
\title{JHU White Matter Tract Atlas}
\usage{
jhu_tracts()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
JHU probabilistic white matter tract atlas with 20 tracts,
thresholded at 25\% probability. Contains 3D mesh geometry
for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
jhu_tracts()
}
\references{
Hua K, et al. (2008). Tract probability maps in
stereotaxic spaces: analyses of white matter anatomy and
tract-specific quantification. \emph{NeuroImage}, 39(1):336-347.
\doi{10.1016/j.neuroimage.2007.07.053}
}
\seealso{
Other ggseg_atlases: 
\code{\link{jhu_wm}()}

Other tract_atlases: 
\code{\link{jhu_wm}()}
}
\concept{ggseg_atlases}
\concept{tract_atlases}
