% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_wm.R
\name{jhu_wm}
\alias{jhu_wm}
\title{ICBM-DTI-81 White Matter Atlas}
\usage{
jhu_wm()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
White matter label atlas derived from the ICBM DTI-81
probabilistic atlas of 81 normal subjects, with 48 regions
projected onto the FreeSurfer \code{cvs_avg35_inMNI152} anatomical
grid for cortical-reference context and smoother boundaries.
}
\examples{
jhu_wm()
plot(jhu_wm())
}
\references{
Mori S, et al. (2005). MRI Atlas of Human White
Matter. Elsevier. ISBN: 978-0444517418

Wakana S, et al. (2007). Reproducibility of quantitative
tractography methods applied to cerebral white matter.
\emph{NeuroImage}, 36(3):630-644.
\doi{10.1016/j.neuroimage.2007.02.049}
}
\seealso{
Other ggseg_atlases: 
\code{\link{jhu_tracts}()}

Other tract_atlases: 
\code{\link{jhu_tracts}()}
}
\concept{ggseg_atlases}
\concept{tract_atlases}
